Disable proxy_next_upstream
When proxy_next_upstream is enabled and the a request fails, Nginx
tries to connect to the next upstream. Since there is never a
second upstream in our setup, Nginx will start responding with
502s. This is fine if there a multiple upstreams but since there
is only ever one upstream, it's better to have Nginx just send
further request to the upstream anyway.